Comments, check user exists before delete

+++ app/Database.py
@@ -32,6 +32,16 @@ class Database:
 class UserTable():
 
     def get_users(self, limit=10, offset=0, sort="id:asc"):
+        """Get all users
+
+        Args:
+            limit (int, optional): max number of users to return. Defaults to 10.
+            offset (int, optional): starting index in the list of users to return. Defaults to 0.
+            sort (str, optional): order the list of users with the combination attribute: order. Defaults to "id:asc".
+
+        Returns:
+            list: list of users found
+        """
         attr, order = sort.split(':')
         query = f"""SELECT *
         FROM person
@@ -44,10 +54,23 @@ class UserTable():
 
     @classmethod
     def number_user(cls):
+        """Get number of users
+
+        Returns:
+            int: Number of users
+        """
         with Database.query("SELECT COUNT(*) AS number_user FROM person") as cur:
             return dict(cur.fetchone()).get("number_user", 0)
 
     def get_user_by_id(self, user_id):
+        """Get user by ID
+
+        Args:
+            user_id (int): User ID
+
+        Returns:
+            dict: User found
+        """
         query = f"""SELECT *
             FROM person
             WHERE id=%s;"""
@@ -58,6 +81,14 @@ class UserTable():
             }
 
     def post_user(self, user_data: PartialUser):
+        """Create new users
+
+        Args:
+            user_data (PartialUser): User to create
+
+        Returns:
+            dict: Success operation
+        """
         prepare_data = (
             user_data.first_name, user_data.last_name,
             user_data.gender.value, user_data.email,
@@ -75,6 +106,15 @@ class UserTable():
             return {"success": True}
 
     def patch_user(self, user_id: int, user_data: PartialUser):
+        """Fix some users attributes except his ID
+
+        Args:
+            user_id (int): user ID
+            user_data (PartialUser): new data
+
+        Returns:
+            dict: Success operation
+        """
         prepare_data = (
             user_data.first_name, user_data.last_name,
             user_data.gender.value, user_data.email,
@@ -93,6 +133,15 @@ class UserTable():
                 return {"success": True}
 
     def put_user(self, user_id: int, new_user: User):
+        """Replace user by another
+
+        Args:
+            user_id (int): user to replace, his ID
+            new_user (User): new user
+
+        Returns:
+            dict: Success operation
+        """
         # check if user with chosen id already exists
         if self.get_user_by_id(new_user.id) is None:
             prepare_data = (
@@ -117,6 +166,17 @@ class UserTable():
         }
 
     def delete_user(self, user_id: int):
-        query = "DELETE FROM person WHERE id=%s"
-        with Database.query(query, (user_id,)) as cur:
-            return {"success": True}
+        """Delete a user
+
+        Args:
+            user_id (int): user to delete, his ID
+
+        Returns:
+            dict: Success operation
+        """
+        user_exists = self.get_user_by_id(user_id)
+        if user_exists:
+            query = "DELETE FROM person WHERE id=%s"
+            with Database.query(query, (user_id,)) as cur:
+                return {"success": True}
+        return {"detail": "Invalid user ID: {user_id}"}
